printing.latex: Move duplicated code to a mixin class
This also renames `DefaultPrinting` to have a name that is more consistent.
