+++ sympy/printing/latex.py
@@ -121,6 +121,32 @@ _between_two_numbers_p = (
 )
 
 
+class LatexPrintable:
+    """
+    The default implementation of latex printing for SymPy classes.
+
+    Mix in this trait into a class to get proper latex printing with IPython.
+    """
+
+    # We don't define _repr_png_ here because it would add a large amount of
+    # data to any notebook containing SymPy expressions, without adding
+    # anything useful to the notebook. It can still enabled manually, e.g.,
+    # for the qtconsole, with init_printing().
+    def _repr_latex_(self):
+        """
+        IPython/Jupyter LaTeX printing
+
+        To change the behavior of this (e.g., pass in some settings to LaTeX),
+        use init_printing(). init_printing() will also enable LaTeX printing
+        for built in numeric types like ints and container types that contain
+        SymPy objects, like lists and dictionaries of expressions.
+        """
+        s = latex(self, mode='plain')
+        return "$\\displaystyle %s$" % s
+
+    _repr_latex_orig = _repr_latex_
+
+
 class LatexPrinter(Printer):
     printmethod = "_latex"
