+++ sympy/printing/str.py
@@ -16,6 +16,26 @@ from mpmath.libmp import prec_to_dps, to_str as mlib_to_str
 from sympy.utilities import default_sort_key
 
 
+class StrPrintable:
+    """
+    The default implementation of printing for SymPy classes.
+
+    This implements a hack that allows us to print elements of built-in
+    Python containers in a readable way. Natively Python uses ``repr()``
+    even if ``str()`` was explicitly requested. Mix in this trait into
+    a class to get proper default printing.
+
+    """
+
+    # Note, we always use the default ordering (lex) in __str__ and __repr__,
+    # regardless of the global setting. See issue 5487.
+    def __str__(self):
+        from sympy.printing.str import sstr
+        return sstr(self, order=None)
+
+    __repr__ = __str__
+
+
 class StrPrinter(Printer):
     printmethod = "_sympystr"
     _default_settings = {
